requirements.txt
firebase_admin==6.6.0
pyperclip==1.9.0
python-telegram-bot==21.10
